requirements.txt
Flask==3.1.2
requests==2.32.5
qrcode==8.2
Pillow==11.3.0
numpy==2.3.3
torch==2.3.1
torchvision==0.18.1
scikit-image==0.22.5
gunicorn==23.0.0
werkzeug==3.1.3
